# docs/api/commerceflow-openapi.yaml
...
openapi: 3.0.3
info:
  title: CommerceFlow API
  version: 1.0.0
  description: |
    Microservices-based e-commerce order processing.

    Placing an order is **asynchronous**. `POST /api/orders` returns `201` with status `CREATED`
    as soon as the order is accepted and the orchestrator has been started; stock
    reservation, payment and notification then run over Kafka. Poll `GET /api/orders/{id}` to
    follow it — the status walks
    `CREATED → INVENTORY_RESERVED → PAID → COMPLETED`, or moves to `CANCELLED` with
    `failedStep` and `failureReason` telling you where and why it stopped.

    Every response is wrapped in a uniform envelope: `ApiResponse` on success, `ErrorResponse`
    on failure. Both carry the `correlationId`, which is the id to quote in a support request —
    it stitches the whole saga together in the logs.
  contact:
    name: CommerceFlow Platform Team
    email: platform@commerceflow.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local API Gateway (docker compose)
  - url: https://api.commerceflow.io
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new customer account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409":
          description: The address is already registered (`EMAIL_ALREADY_REGISTERED`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Exchange credentials for an access and refresh token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Token pair issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: |
            `INVALID_CREDENTIALS`. Returned identically for an unknown address and a wrong
            password, so the endpoint cannot be used to enumerate registered customers.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: The account is disabled (`ACCOUNT_DISABLED`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a refresh token into a fresh token pair
      description: |
        Refresh tokens are single use. Presenting one that has already been rotated is treated
        as a compromise indicator and revokes **every** live session of the account.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: A new token pair; the presented token is now revoked
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: "`REFRESH_TOKEN_INVALID` or `TOKEN_EXPIRED`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current access token and every refresh token of the account
      description: |
        The access token is added to a Redis deny-list for the remainder of its natural life,
        which is the only way to end a self-contained token's session early.
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Return the account behind the current access token
      responses:
        "200":
          description: The account
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request a password reset link
      description: |
        Answers `202` for a registered address, an unregistered one and a disabled account
        alike. An endpoint that responded differently would let anyone check which addresses
        have accounts, so the message says *if an account exists* rather than claiming an email
        was sent.

        The emailed token is single use and expires in 30 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: ada@commerceflow.io }
      responses:
        "202":
          description: Request accepted, whether or not an account exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /api/auth/reset-password:
    post:
      tags: [Authentication]
      summary: Set a new password using an emailed token
      description: |
        Single use. Completing a reset **revokes every refresh token the account holds** — a
        reset is what someone does when they believe their account is compromised, and leaving
        the intruder's session alive would defeat the point. It also marks the address verified,
        since redeeming the token proves the mailbox was reachable.

        Unknown, expired and already-used tokens all return the same `400`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, description: The token from the emailed link }
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: At least 8 characters with one letter and one digit
                  example: N3w-pass-2026
      responses:
        "200":
          description: Password changed; every session was signed out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "400":
          description: The link is unknown, already used, or expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/auth/verify-email:
    post:
      tags: [Authentication]
      summary: Confirm an email address
      description: |
        Redeems the token sent at registration. Clicking twice is not an error.

        Whether an unverified account may sign in is a deployment switch
        (`commerceflow.auth.require-verified-email`), off in the demo stack and expected to be
        on in production.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: The token from the emailed link }
      responses:
        "200":
          description: Address confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "400":
          description: The link is unknown, already used, or expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/auth/resend-verification:
    post:
      tags: [Authentication]
      summary: Ask for another confirmation link
      description: |
        Answers identically for an unknown address, an already-verified one and one that needs
        confirming, for the same reason as `forgot-password`. Issuing a new link invalidates any
        previous one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: ada@commerceflow.io }
      responses:
        "202":
          description: Request accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }

  # ------------------------------------------------------------------ account administration
...
  /api/auth/guest:
    post:
      tags: [Authentication]
      summary: Check out without an account
      description: |
        Creates an account with **no password** and returns an ordinary session for it. A guest is
        a real customer as far as everything downstream is concerned — the basket, the order, the
        payment and the saga are unchanged.

        They cannot sign in until they set a password through the ordinary reset flow, at which
        point their order history is already there.

        **An address that already has an account is refused with `409`.** This endpoint hands out a
        session in exchange for an email address and nothing else; returning one for an existing
        account would let anybody type your address and be signed in as you.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                fullName: { type: string, maxLength: 150 }
      responses:
        "201": { description: A session }
        "409": { $ref: "#/components/responses/Conflict" }

  # ------------------------------------------------------------------ payments
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by `POST /api/auth/login`

  # This file is the hand-written contract for the paths that matter most. It is deliberately
  # not exhaustive: each service serves its own generated definition at /v3/api-docs, aggregated
  # behind the gateway at /swagger-ui.html, and that is authoritative for everything else --
  # notably the back-office endpoints for warehouses, stock levels, product images and review
  # moderation, which no external client calls.
  responses:
    Unauthorized:
      description: No valid access token (`UNAUTHORIZED`, `TOKEN_EXPIRED`, `TOKEN_INVALID`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ValidationFailed:
      description: Request validation failed; `fieldErrors` lists every violated constraint
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: |
        The request is well formed but the resource is not in a state that allows it
        (`ORDER_NOT_MODIFIABLE`, `CONFLICT`). The message says what to do about it.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Too many requests (`RATE_LIMITED`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ApiResponse:
      type: object
      description: Uniform success envelope.
      properties:
        success: { type: boolean, example: true }
        data: { description: The payload }
        message: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        correlationId:
          type: string
          nullable: true
          description: Quote this in a support request; it stitches the saga together in the logs.
    ErrorResponse:
      type: object
      description: Uniform error envelope.
      properties:
        success: { type: boolean, example: false }
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        code:
          type: string
          example: INSUFFICIENT_STOCK
          description: Machine readable error code; stable across releases.
        message: { type: string }
        path: { type: string, example: /api/orders }
        correlationId: { type: string, nullable: true }
        fieldErrors:
          type: array
          nullable: true
          items:
            type: object
            properties:
              field: { type: string, example: email }
              rejectedValue: {}
              message: { type: string, example: must be a valid address }
...
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email: { type: string, format: email, maxLength: 255, example: ada@commerceflow.io }
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: At least one letter and one digit.
          example: S3cret-pass
        fullName: { type: string, maxLength: 150, example: Ada Lovelace }
        phone: { type: string, maxLength: 32, nullable: true, example: "+31612345678" }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: ada@commerceflow.io }
        password: { type: string, format: password, example: S3cret-pass }
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: "Short lived JWT; send as the `Authorization: Bearer <token>` header."
        refreshToken:
          type: string
          description: Long lived and single use — presenting it rotates the pair.
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, format: int64, example: 900, description: Seconds }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        phone: { type: string, nullable: true }
        roles:
          type: array
          items: { type: string, enum: [CUSTOMER, ADMIN] }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
...